+++ tests/lightning_base/test_device.py
@@ -16,6 +16,8 @@ This module contains unit tests for new device API Lightning classes.
 """
 # pylint: disable=too-many-arguments, unused-argument
 
+import itertools
+
 import numpy as np
 import pennylane as qml
 import pytest
@@ -99,10 +101,12 @@ else:
 if not LightningDevice._CPP_BINARY_AVAILABLE:  # pylint: disable=protected-access
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)
 
+fixture_params = itertools.product([3, None], [np.complex64, np.complex128])  # wires x c_dtype
 
-@pytest.fixture(params=[np.complex64, np.complex128])
+
+@pytest.fixture(params=fixture_params)
 def dev(request):
-    return LightningDevice(wires=3, c_dtype=request.param)
+    return LightningDevice(wires=request.param[0], c_dtype=request.param[1])
 
 
 class TestHelpers:
@@ -203,6 +207,121 @@ class TestHelpers:
     def test_supports_adjoint(self, circuit, expected):
         """Test that _supports_adjoint returns the correct boolean value."""
         assert _supports_adjoint(circuit) == expected
+
+    @pytest.mark.skipif(
+        device_name == "lightning.tensor", reason="lightning.tensor does not contain a state vector"
+    )
+    @pytest.mark.parametrize("device_wires", [None, 2])
+    def test_state_vector_init(self, device_wires):
+        """Test that the state-vector is not created during initialization"""
+        dev = LightningDevice(wires=device_wires)
+        assert dev._statevector == None
+
+    @pytest.mark.parametrize(
+        "circuit_in, n_wires, expected_circuit_out",
+        [
+            (
+                QuantumScript(
+                    [
+                        qml.RX(0.1, 0),
+                        qml.CNOT([1, 0]),
+                        qml.RZ(0.1, 1),
+                        qml.CNOT([2, 1]),
+                    ],
+                    [qml.expval(qml.Z(0))],
+                ),
+                3,
+                QuantumScript(
+                    [
+                        qml.RX(0.1, 0),
+                        qml.CNOT([1, 0]),
+                        qml.RZ(0.1, 1),
+                        qml.CNOT([2, 1]),
+                    ],
+                    [qml.expval(qml.Z(0))],
+                ),
+            ),
+            (
+                QuantumScript(
+                    [
+                        qml.RX(0.1, 0),
+                        qml.CNOT([1, 4]),
+                        qml.RZ(0.1, 4),
+                        qml.CNOT([2, 1]),
+                    ],
+                    [qml.expval(qml.Z(6))],
+                ),
+                5,
+                QuantumScript(
+                    [
+                        qml.RX(0.1, 0),
+                        qml.CNOT([1, 2]),
+                        qml.RZ(0.1, 2),
+                        qml.CNOT([3, 1]),
+                    ],
+                    [qml.expval(qml.Z(4))],
+                ),
+            ),
+        ],
+    )
+    def test_dynamic_wires_from_circuit(self, circuit_in, n_wires, expected_circuit_out):
+        """Test that dynamic_wires_from_circuit returns correct circuit and creates state-vectors properly"""
+        device = LightningDevice(wires=None)
+        circuit_out = device.dynamic_wires_from_circuit(circuit_in)
+
+        assert circuit_out.num_wires == n_wires
+        assert circuit_out.wires == qml.wires.Wires(range(n_wires))
+        assert circuit_out.operations == expected_circuit_out.operations
+        assert circuit_out.measurements == expected_circuit_out.measurements
+
+        if device_name != "lightning.tensor":
+            assert device._statevector._num_wires == n_wires
+            assert device._statevector._wires == qml.wires.Wires(range(n_wires))
+
+    @pytest.mark.parametrize(
+        "circuit_in, n_wires, wires_list",
+        [
+            (
+                QuantumScript(
+                    [
+                        qml.RX(0.1, 0),
+                        qml.CNOT([1, 0]),
+                        qml.RZ(0.1, 1),
+                        qml.CNOT([2, 1]),
+                    ],
+                    [qml.expval(qml.Z(0))],
+                ),
+                3,
+                [0, 1, 2],
+            ),
+            (
+                QuantumScript(
+                    [
+                        qml.RX(0.1, 0),
+                        qml.CNOT([1, 4]),
+                        qml.RZ(0.1, 4),
+                        qml.CNOT([2, 1]),
+                    ],
+                    [qml.expval(qml.Z(6))],
+                ),
+                7,
+                [0, 1, 4, 2, 6],
+            ),
+        ],
+    )
+    def test_dynamic_wires_from_circuit_fixed_wires(self, circuit_in, n_wires, wires_list):
+        """Test that dynamic_wires_from_circuit does not alter the circuit if wires are fixed and state-vector is created properly"""
+        device = LightningDevice(wires=n_wires)
+        circuit_out = device.dynamic_wires_from_circuit(circuit_in)
+
+        assert circuit_out.num_wires == circuit_in.num_wires
+        assert circuit_out.wires == qml.wires.Wires(wires_list)
+        assert circuit_out.operations == circuit_in.operations
+        assert circuit_out.measurements == circuit_in.measurements
+
+        if device_name != "lightning.tensor":
+            assert device._statevector._num_wires == n_wires
+            assert device._statevector._wires == qml.wires.Wires(range(n_wires))
 
 
 class TestInitialization:
@@ -554,6 +673,46 @@ class TestExecution:
         for r, e in zip(res, expected):
             assert np.allclose(r, e)
 
+    def test_execute_tape_batch_with_dynamic_wires(self):
+        """Test that execute handles multiple tapes with dynamic number of wires."""
+
+        qs0 = QuantumScript(
+            [
+                qml.RX(0.1, 0),
+                qml.CNOT([1, 0]),
+                qml.RZ(0.1, 1),
+                qml.CNOT([2, 1]),
+            ],
+            [qml.state()],
+        )
+        qs1 = QuantumScript(
+            [
+                qml.RX(0.1, 0),
+                qml.CNOT([1, 0]),
+                qml.RZ(0.1, 1),
+                qml.CNOT([0, 1]),
+            ],
+            [qml.state()],
+        )
+        qs2 = QuantumScript(
+            [
+                qml.RX(0.1, 4),
+                qml.CNOT([2, 4]),
+                qml.RZ(0.1, 2),
+                qml.CNOT([1, 2]),
+                qml.CNOT([0, 2]),
+            ],
+            [qml.state()],
+        )
+        dev = LightningDevice(wires=None)
+        result = dev.execute([qs0, qs1, qs2])
+
+        dev_ref = DefaultQubit(max_workers=1)
+        result_ref = dev_ref.execute([qs0, qs1, qs2])
+
+        for r, e in zip(result, result_ref):
+            assert np.allclose(r, e)
+
     @pytest.mark.parametrize("phi, theta", list(zip(PHI, THETA)))
     @pytest.mark.parametrize("wires", (["a", "b", -3], [0, "target", "other_target"]))
     def test_custom_wires_execute(self, phi, theta, wires):
@@ -895,12 +1054,14 @@ class TestDerivatives:
         with pytest.raises(qml.DeviceError, match="Finite shots are not supported"):
             _, _ = program([qs])
 
+    @pytest.mark.parametrize("device_wires", [None, 4])
     @pytest.mark.parametrize("phi", PHI)
     @pytest.mark.parametrize("execute_and_derivatives", [True, False])
-    def test_derivatives_tape_batch(self, phi, execute_and_derivatives, batch_obs):
+    def test_derivatives_tape_batch(self, device_wires, phi, execute_and_derivatives, batch_obs):
         """Test that results are correct when we execute and compute derivatives for a batch of
-        tapes."""
-        device = LightningDevice(wires=4, batch_obs=batch_obs)
+        tapes with and without dynamic wires."""
+
+        device = LightningDevice(wires=device_wires, batch_obs=batch_obs)
 
         ops = [qml.X(0), qml.X(1)]
         if device_name == "lightning.qubit":
@@ -1258,12 +1419,14 @@ class TestVJP:
         ):
             _ = dev.execute_and_compute_vjp(qs, dy, config)
 
+    @pytest.mark.parametrize("device_wires", [None, 4])
     @pytest.mark.parametrize("phi", PHI)
     @pytest.mark.parametrize("execute_and_derivatives", [True, False])
-    def test_vjp_tape_batch(self, phi, execute_and_derivatives, batch_obs):
-        """Test that results are correct when we execute and compute derivatives for a batch of
-        tapes."""
-        device = LightningDevice(wires=4, batch_obs=batch_obs)
+    def test_vjp_tape_batch(self, device_wires, phi, execute_and_derivatives, batch_obs):
+        """Test that results are correct when we execute and compute vjp for a batch of
+        tapes with and without dynamic wires."""
+
+        device = LightningDevice(wires=device_wires, batch_obs=batch_obs)
 
         ops = [
             qml.X(0),