+++ tests/test_vjp.py
@@ -14,6 +14,7 @@
 """
 Tests for the ``vjp`` method.
 """
+import itertools
 import math
 
 import pennylane as qml
@@ -32,9 +33,11 @@ if device_name == "lightning.tensor":
 class TestVectorJacobianProduct:
     """Tests for the `vjp` function"""
 
-    @pytest.fixture(params=[np.complex64, np.complex128])
+    fixture_params = itertools.product([np.complex64, np.complex128], [None, 2])
+
+    @pytest.fixture(params=fixture_params)
     def dev(self, request):
-        return qml.device(device_name, wires=2, c_dtype=request.param)
+        return qml.device(device_name, wires=request.param[1], c_dtype=request.param[0])
 
     def test_multiple_measurements(self, tol, dev):
         """Tests provides correct answer when provided multiple measurements."""
@@ -261,9 +264,11 @@ class TestVectorJacobianProduct:
 class TestBatchVectorJacobianProduct:
     """Tests for the batch_vjp function"""
 
-    @pytest.fixture(params=[np.complex64, np.complex128])
+    fixture_params = itertools.product([np.complex64, np.complex128], [None, 2])
+
+    @pytest.fixture(params=fixture_params)
     def dev(self, request):
-        return qml.device(device_name, wires=2, c_dtype=request.param)
+        return qml.device(device_name, wires=request.param[1], c_dtype=request.param[0])
 
     def test_one_tape_no_trainable_parameters_1(self, dev):
         """A tape with no trainable parameters will simply return None"""